requirements.txt
streamlit
watchdog
langchain
langchain-community
langchain-openai
langgraph
chromadb
pydantic
python-dotenv

sympy

paddleocr
pillow
easyocr

faster-whisper

sentence-transformers
faiss-cpu